implementing new style
